# docs/07-data-api-events/specs/openapi.platform.yaml
openapi: 3.1.2
info:
  title: Korean Equity RWA PoC Platform API
  version: 1.0.0
  description: 합성 데이터만 사용하는 7단계 플랫폼 HTTP 계약이다. 상태변경 명령의 202 응답은 업무 완료가 아니다.
servers:
  - url: http://localhost:3000/api/v1
security:
  - syntheticBearer: []
...
paths:
...
  /products:
    get:
      tags: [CommonQueries]
      operationId: listProducts
      summary: 상품목록과 기능별 가능상태 조회
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 상품목록과 기능별 가능상태
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedProducts' }
  /products/{securityId}:
    get:
      tags: [CommonQueries]
      operationId: getProduct
      summary: 종목별 상품 상세 조회
      parameters:
        - $ref: '#/components/parameters/SecurityId'
      responses:
        '200':
          description: 상품 상세
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404': { $ref: '#/components/responses/Error' }
...
  /workflows/{workflowId}:
    get:
      tags: [CommonQueries]
      operationId: getWorkflow
      summary: 업무의 독립 상태축 조회
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: 독립 상태축과 투영 완전성
          content:
            application/json:
              schema:
                $ref: './schemas/domain.schema.json#/$defs/WorkflowView'
        '404': { $ref: '#/components/responses/Error' }
...
  /disclosures/current:
    get:
      tags: [CommonQueries]
      operationId: getCurrentDisclosure
      summary: 현재 적용 위험공시와 책임기관 조회
      responses:
        '200':
          description: 현재 적용 위험공시
          content:
            application/json:
              schema: { $ref: './schemas/domain.schema.json#/$defs/Disclosure' }
        '404': { $ref: '#/components/responses/Error' }
...
  /disclosure-consents:
    post:
      tags: [InvestorCommands]
      operationId: createDisclosureConsent
      summary: 현재 위험공시에 전자 동의 제출
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/domain.schema.json#/$defs/DisclosureConsentRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /complaints:
    get:
      tags: [CommonQueries]
      operationId: listComplaints
      summary: 본인 또는 역할 범위의 민원 목록 조회
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 민원 목록
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, projection]
                properties:
                  items:
                    type: array
                    items: { $ref: './schemas/domain.schema.json#/$defs/ComplaintView' }
                  nextCursor: { type: string }
                  projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
    post:
      tags: [InvestorCommands]
      operationId: createComplaint
      summary: 투자자 민원 접수
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: './schemas/domain.schema.json#/$defs/ComplaintRequest' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /complaints/{complaintId}:
    get:
      tags: [CommonQueries]
      operationId: getComplaint
      summary: 본인 또는 담당 범위의 민원 상세 조회
      parameters:
        - $ref: '#/components/parameters/ComplaintId'
      responses:
        '200':
          description: 책임기관, 답변, 정정과 종결 근거를 포함한 민원
          content:
            application/json:
              schema: { $ref: './schemas/domain.schema.json#/$defs/ComplaintView' }
        '404': { $ref: '#/components/responses/Error' }
  /wallet-link-requests:
    post:
      tags: [InvestorCommands]
      operationId: requestWalletLink
      summary: 고객 전용 지갑 연결 요청
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [wallet, ownershipSignature]
              properties:
                wallet: { $ref: './schemas/common.schema.json#/$defs/EvmAddress' }
                ownershipSignature: { type: string, pattern: '^0x[0-9a-fA-F]{130}$' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /wallet-replacement-requests:
    post:
      tags: [InvestorCommands]
      operationId: requestWalletReplacement
      summary: 고객 전용 지갑 교체 요청
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [oldWallet, newWallet, reasonKo, newWalletSignature]
              properties:
                oldWallet: { $ref: './schemas/common.schema.json#/$defs/EvmAddress' }
                newWallet: { $ref: './schemas/common.schema.json#/$defs/EvmAddress' }
                reasonKo: { type: string, minLength: 1, maxLength: 500 }
                newWalletSignature: { type: string, pattern: '^0x[0-9a-fA-F]{130}$' }
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /institution/tasks/{taskId}/decisions:
    post:
      tags: [InstitutionCommands]
      operationId: decideInstitutionTask
      summary: 기관 업무 승인 또는 거절 기록
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas/domain.schema.json#/$defs/InstitutionDecision'
      responses:
        '202': { $ref: '#/components/responses/Accepted' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    syntheticBearer:
      type: http
      scheme: bearer
      bearerFormat: Synthetic-JWT
    adapterSignature:
      type: apiKey
      in: header
      name: X-Adapter-Key-Id
      description: 본문 안의 Ed25519 signature와 함께 검증한다.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 512 }
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ComplaintId:
      name: complaintId
      in: path
      required: true
      schema: { type: string, format: uuid }
    SecurityId:
      name: securityId
      in: path
      required: true
      schema: { $ref: './schemas/common.schema.json#/$defs/SecurityId' }
...
  responses:
...
    Accepted:
      description: 비동기 업무 접수
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/AcceptedCommand' }
    Error:
      description: 안정적인 기계 오류와 한국어 다음 행동
      content:
        application/json:
          schema: { $ref: './schemas/common.schema.json#/$defs/ErrorResponse' }
...
  schemas:
...
    Product:
      type: object
      additionalProperties: false
      required: [securityId, nameKo, referenceVersion, candidateStatus, representative, availability, blockingReasons, notices, simulation, projection]
      properties:
        securityId: { $ref: './schemas/common.schema.json#/$defs/SecurityId' }
        isin: { $ref: './schemas/common.schema.json#/$defs/Isin' }
        nameKo: { type: string, minLength: 1 }
        referenceVersion: { type: string, minLength: 1 }
        tokenAddress: { $ref: './schemas/common.schema.json#/$defs/EvmAddress' }
        candidateStatus: { type: string, enum: [CANDIDATE, REVIEWED, INFORMATION_UNCONFIRMED] }
        representative: { type: boolean }
        availability:
          type: object
          additionalProperties: false
          required: [primary, secondary, redemption]
          properties:
            primary: { type: string, enum: [ENABLED, DISABLED] }
            secondary: { type: string, enum: [ENABLED, DISABLED] }
            redemption: { type: string, enum: [ENABLED, DISABLED, MANUAL_REVIEW] }
        blockingReasons:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [code, messageKo]
            properties:
              code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$' }
              messageKo: { type: string, minLength: 1 }
        notices:
          type: object
          additionalProperties: false
          required: [rightsNatureKo, custodyRiskKo, transferRestrictionKo, settlementKo, dividendKo, votingKo, redemptionKo]
          properties:
            rightsNatureKo: { type: string, minLength: 1 }
            custodyRiskKo: { type: string, minLength: 1 }
            transferRestrictionKo: { type: string, minLength: 1 }
            settlementKo: { type: string, minLength: 1 }
            dividendKo: { type: string, minLength: 1 }
            votingKo: { type: string, minLength: 1 }
            redemptionKo: { type: string, minLength: 1 }
        simulation: { const: true }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
    PagedProducts:
      type: object
      additionalProperties: false
      required: [items, projection]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Product' } }
        nextCursor: { type: string }
        projection: { $ref: './schemas/common.schema.json#/$defs/ProjectionMetadata' }
...